requirements.txt
streamlit==1.32.0
torch==2.10.0
torchvision==0.20.0
numpy==1.26.4
Pillow==10.2.0
opencv-python==4.9.0.80
matplotlib==3.8.3
pandas==2.2.1
scikit-learn==1.4.1.post1